FDG-9818 Update report display names
