+++ src/components/combo-select/combo-currency-select/combo-select-dropdown/combo-select-dropdown.jsx
@@ -73,7 +73,9 @@ const ComboSelectDropdown = ({
   };
 
   const clearFilter = () => {
-    changeHandler(null);
+    if (changeHandler) {
+      changeHandler(null);
+    }
     // fire artificial event to reset field
     onFilterChange({
       target: {
