+++ src/components/published-reports/download-report-table/download-report-table-row/download-report-table-row.tsx
@@ -17,7 +17,7 @@ import {
 } from './download-report-table-row.module.scss';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCloudArrowDown, faCircleCheck } from '@fortawesome/free-solid-svg-icons';
-import { getFileTypeImage, splitFileName } from '../../util/util';
+import { getFileDisplay, getFileTypeImage } from '../../util/util';
 import { IReports } from '../../reports-section/reports-section';
 import { getDateLabelForReport } from '../../../../helpers/dataset-detail/report-helpers';
 import { getFileSize } from '../../download-report/download-helpers';
@@ -50,21 +50,11 @@ const DownloadReportTableRow: FunctionComponent<{ reportFile: IReports; isDailyR
           setFileSize(size);
         });
       }
-      // grab the file extension
-      const groupName = curReportFile.report_group_desc;
-      const splitReportPath = curReportFile.path.split('.');
-      if (splitReportPath.length > 0) {
-        const reportFileType = splitReportPath[splitReportPath.length - 1];
-        const apiFileType = '(.' + reportFileType + ')';
-        const downloadFileType = '.' + reportFileType;
-        // Remove parenthesis from file name -> ex. fileName (.pdf) to fileName.pdf
-        const fullDisplayName = groupName.replace(' ' + apiFileType, downloadFileType);
-        //Split file name so overflow ellipsis can be used in the middle of the name
-        const fileDisplayName = splitFileName(fullDisplayName, fullDisplayName.length - 8);
-        setDisplayName(fileDisplayName || '');
-        setFileType(downloadFileType);
-        setFileTypeImage(getFileTypeImage(downloadFileType));
-      }
+
+      const fileDisplay = getFileDisplay(curReportFile);
+      setDisplayName(fileDisplay.displayName);
+      setFileType(fileDisplay.fileType);
+      setFileTypeImage(getFileTypeImage(fileDisplay.fileType));
     }
   };
